backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openai
openpyxl
python-dotenv
